+++ packages/client/src/pages/game/components/GamePage/useGamePageData.ts
@@ -1,10 +1,7 @@
 import { RefObject, useCallback, useEffect, useRef, useState } from 'react'
 import { Game, type GameEvent } from '@/game/engine/Game'
-import { useNotification } from '@/hooks/useNotification'
 import { modalEvents } from './GamePage.constants'
-
-const AWAY_DELAY_MS = 3000
-const REMINDER_DELAY_MS = 60000
+import { useGameNotifications } from './useGameNotifications'
 
 type UseGamePageDataProps = {
   canvasRef: RefObject<HTMLCanvasElement | null>
@@ -49,57 +46,7 @@ export const useGamePageData = ({ canvasRef }: UseGamePageDataProps) => {
     return gameRef.current?.getSnapshot() ?? null
   }, [])
 
-  const { show: showNotification } = useNotification()
-  const showNotificationRef = useRef(showNotification)
-  showNotificationRef.current = showNotification
-
-  useEffect(() => {
-    let alertId: ReturnType<typeof setTimeout> | null = null
-    let reminderId: ReturnType<typeof setTimeout> | null = null
-
-    const clearTimers = () => {
-      if (alertId) {
-        clearTimeout(alertId)
-        alertId = null
-      }
-      if (reminderId) {
-        clearTimeout(reminderId)
-        reminderId = null
-      }
-    }
-
-    const handleVisibility = () => {
-      if (document.visibilityState === 'hidden') {
-        const snap = getSnapshot()
-        if (snap?.state !== 'playing') return
-
-        alertId = setTimeout(() => {
-          const current = getSnapshot()
-          if (current?.state === 'playing') {
-            showNotificationRef.current('Вы под обстрелом!', {
-              body: `Ваш корабль под огнём. Жизней: ${current.lives}`,
-              tag: 'game-alert',
-            })
-          }
-        }, AWAY_DELAY_MS)
-
-        reminderId = setTimeout(() => {
-          showNotificationRef.current('Вы отсутствуете минуту', {
-            body: 'Корабль всё ещё под обстрелом! Нажмите, чтобы вернуться.',
-            tag: 'game-reminder',
-          })
-        }, REMINDER_DELAY_MS)
-      } else {
-        clearTimers()
-      }
-    }
-
-    document.addEventListener('visibilitychange', handleVisibility)
-    return () => {
-      document.removeEventListener('visibilitychange', handleVisibility)
-      clearTimers()
-    }
-  }, [getSnapshot])
+  useGameNotifications(getSnapshot)
 
   const initGame = async () => {
     setIsLoading(true)
